feat(synflat): add SyntheticFlatGenerator, SyntheticFlatStep, SyntheticFlatPanel and full project persistence (F-SynFlat)
- astroai/processing/flat/: SyntheticFlatGenerator + SyntheticFlatStep (neue Pipeline-Stufe)
- astroai/ui/widgets/synthetic_flat_panel.py: SyntheticFlatPanel UI-Dock
- 9 Quelldateien: mypy-Typ-Fixes (explicit casts, NDArray, dict generics)
- Bug-Fix: calibrator.py:275 _catalog_to_pixels 2-tuple → 3-tuple
- project_file.py, app.py, models.py: vollständige bidirektionale Projekt-Persistenz

Status: 1900 Tests, 0 mypy-Fehler, 0 ruff-Fehler

+++ astroai/engine/photometry/aperture.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import numpy as np
 from numpy.typing import NDArray
-from scipy import ndimage
 
 
 class AperturePhotometry:
@@ -69,5 +68,5 @@ class AperturePhotometry:
         half_max = peak / 2.0
         above = stamp >= half_max
         n_pixels = float(np.sum(above))
-        fwhm = 2.0 * np.sqrt(n_pixels / np.pi)
+        fwhm = float(2.0 * np.sqrt(n_pixels / np.pi))
         return max(fwhm, 1.5)
